Refactor: 공통 필드를 Common으로 분리
- id, createdAt, updatedAt 필드를 Common로 이동
- User, Channel, Message가 Common을 상속하도록 변경
- 중복 필드 제거하여 코드 중복 최소화
