+++ src/main/java/com/sprint/mission/discodeit/entity/User.java
@@ -2,11 +2,7 @@ package com.sprint.mission.discodeit.entity;
 
 import java.util.UUID;
 
-public class User {
-    private UUID id;
-    private Long createdAt;
-    private Long updatedAt;
-
+public class User extends Common{
     // 아이디 - 디스코드에서는 이메일 사용
     private String username;
     // 비밀번호
@@ -15,26 +11,11 @@ public class User {
     private String nickname;
 
     public User(String username, String password, String nickname) {
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = createdAt; // 초기엔 생성만 하므로 createdAt과 같다.
-
         this.username = username;
         this.password = password;
         this.nickname = nickname;
     }
 
-    public UUID getId() {
-        return id;
-    }
-
-    public Long getCreatedAt() {
-        return createdAt;
-    }
-
-    public Long getUpdatedAt() {
-        return updatedAt;
-    }
 
     public String getUsername() {
         return username;
@@ -48,13 +29,7 @@ public class User {
         return nickname;
     }
 
-    // id, createdAt은 객체 생성 시점에 정해지므로 update 필요 X
-
-    public void updateUpdatedAt(Long updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    // username 또한, 아이디 변경이 불가능하므로 update 필요 X
+    // username은 아이디 변경이 불가능하므로 update 필요 X
 
     public void updatePassword(String password) {
         this.password = password;
